[TTLG] Update feterated Prometheus helm charts
This updates the Prometheus configurations inside the values of the Helm charts
so that each instance will set a proper cluster label to be recognized in the
ctlplane interface.
Note that maybe the updated on the ctlplane Helm chart are not strictly needed
and after some tests will be removed.

Signed-off-by: Raoul Scarazzini <rasca@mmul.it>