[TTLG] Make Prometheus 3 federates/1 collector
This commit changes the Prometheus setup so that it will be splitted in four
instances: on ctlplane we will have a federated instance and a main one
that will collect all the federated ones (including the federated ctlplane
itself).

Note that the graph schema still needs to be updated.

It also changes the helm script to be idempotent and able to assign fixed IP
addresses on the various exposed services.

Signed-off-by: Raoul Scarazzini <rasca@mmul.it>